fixType

+++ src/main/java/com/ckb/explorer/domain/dto/UdtH24TransactionsCountDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UdtH24TransactionsCountDto {
 
-    private Integer typeScriptId;
+    private Long typeScriptId;
 
     private Long h24CkbTransactionsCount;
 }
